backend/requirements.txt
flask==2.3.3
flask==2.3.3
flask-sqlalchemy==3.0.2
flask-cors==4.0.0
flask-mail==0.9.1
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==2.3.7
